Replace regex to find links with a parser
Uses https://github.com/gjtorikian/commonmarker to create an AST,
manipulates the AST to rewrite links, and then converts back into
markdown.

Fixes https://github.com/benbalter/jekyll-relative-links/issues/61

Fixing https://github.com/benbalter/jekyll-relative-links/issues/61 was
the main motivation for doing this fundamental change, as it wasn't
obvious how to manipulate the regex to accomplish allowing newlines the
markdown link text (not the URL).

Unforununately the test suite did not respond well to this change, as
much of the suite is expecting that the *only* thing modified is the
link text, whereas since we roundtrip now from source -> ast ->
markdown, there are several bits of the original source that are not
respected. For example, in certain situations ' quotes are replaced with
". There are also some minor newline changes, in fact links with
newlines in the text are rewritten to be on a single line.

Lastly, the handing of footnotes is different, as the AST/parser does
not represent a footnote 'source' as it's own entity, the footnotes are
simply 'followed' and reference links are rewritten to be replaced with
the reference's link.

These changes required quite a bit of surgery on the test suite.